refactor: intent key 상수
