+++ app/src/main/java/woowacourse/movie/model/ReservationResultUiModel.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view
+package woowacourse.movie.model
 
 data class ReservationResultUiModel(
     val title: String,
