+++ app/src/main/java/woowacourse/movie/view/ReservationResultView.kt
@@ -1,0 +1,7 @@
+package woowacourse.movie.view
+
+import woowacourse.movie.model.ReservationResultUiModel
+
+interface ReservationResultView {
+    fun showResult(reservationResult: ReservationResultUiModel)
+}
